requirements.txt
alembic==1.3.3
bleach==3.1.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Markdown==0.3
flask-markdown2==0.0.3
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-QiniuStorage==0.9.5
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
jinja2-markdown==0.0.3
Mako==1.1.1
Markdown==3.2.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
Pillow==7.0.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-editor==1.0.4
qiniu==7.2.6
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.16.1
WTForms==2.2.1
